FindAllString()をanalysisGeneral.tsに移動 #1575
