+++ src/components/makeComList.vue
@@ -36,11 +36,10 @@ import {
   loadProfile,
   writeGlobalSetting,
   getPlatform,
-  FindAllString,
   CreateCopyDateList,
 } from '@/utils/analysisData'
 import { DEFAULT_KYARA_SETTING_UUID } from '@/data/data'
-import { createDefoKyaraDateList, createNewDateList } from '@/utils/analysisGeneral'
+import { createDefoKyaraDateList, createNewDateList, FindAllString } from '@/utils/analysisGeneral'
 import type { deleteDialogRefType } from 'src/components/accessories/deleteDialog.vue'
 import deleteDialog from '@/components/accessories/deleteDialog.vue'
 import DisplaySelectFileView from '@/components/accessories/DisplaySelectFileView.vue'
